Bow works in wasm
confirmed by test vectors
