+++ js/test.mjs
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import { testVectors } from './lib/rainstorm.mjs';
+import { testVectors } from './lib/api.mjs';
 
 testVectors();
